Make a few amendments to the Minder section in the ST-WG report
This specifies that Minder does more than just repositories which might
make it sound confusing in comparison to other projects. Instead it
encompasses several other aspects of the SDLC.

It does not provide a set of recommendations, but instead provides a
framework to build your own rules and provides a set of samples (which
one needs to pick and chose from).

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>